Add animated indicator for active task statuses
Create a shared StatusBadge component with a pulsing dot animation
for running, restarting, and cancelling statuses to make it visually
obvious when a task is actively processing.
